Rename: sections.py -> geometry.py; update references in library
