+++ sectionproperties/pre/geometry.py
